Design changes and bug fixes.
Major design changes to the information at the top of sandbox.

Added text service for sandbox to decouple the text from the presentation.

Added text service for expression operation to decouple the text from the presentation.

Fixed the wrong size of input fields when the screen width is small.

Added PrimeNG buttons to expression actions instead of the custom buttons to move towards a common UI design, as done on the frontpage.

Fixed issue with wrong cursor when the Mirror Mountain visualization is not selectable.

Fixed issue with SVG rectangles in the visualization of Mirror Mountain not opening the expression operation.